ZEN-30735: Support data point tags for "don't send" and "only send" to zing-connector
Added the "noStoreTags" and "noForwardTags" fields to the configuration.yaml file.
Updated the MetricServiceConfiguration and ZingConfiguration classes to handle the new config fields.
Added the appropriate getters methods (methods return the new fields value) to each config class.
Added the hasTagKey method (which returns true if the metric instance has this tag key in its tags collection)
to the Metric class.