ZEN-30735: Support data point tags for "don't send" and "only send" to zing-connector
Added the getStoreMetrics (which returns the batch with all metrics should be stored in the OpenTSDB) to OpenTsdbWriter class.
Added the testGetStoreMetrics to the OpenTsdbWriterTest.java.
Added the "cleanupTags" fields to the configuration.yaml file.
Updated the MetricServiceConfiguration class to handle the new "cleanupTags" config field.
Added the appropriate getter method (the method which returns the value of "cleanupTags" field) to MetricServiceConfiguration class.
Added the "testRemoveMetricTags" test to the OpenTsdbWriterTest.java file.