ZEN-30735: Support data point tags for "don't send" and "only send" to zing-connector
Added the getForwardMetrics (which returns the batch with all metrics should be sent to ZING) to the ZingWriter class.
Added the testGetForwardMetrics test to the ZingWriterTest.java.
Added the "cleanupTags" fields to the configuration.yaml file.
Updated the ZingConfiguration class to handle the new "cleanupTags" config field.
Added the appropriate getter method (the method which returns the value of "cleanupTags" field) to ZingConfiguration class.
Added the "testRemoveMetricTags" test to the ZingWriterTest.java file.